refactor: extract contest priority to priority.ts
